backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
pydantic
python-doctr[torch]
httpx
python-dotenv
rq
redis
openai